requirements.txt
absl-py==0.9.0
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
asgiref==3.2.3
asn1crypto==0.22.0
astor==0.8.1
astroid==1.5.3
astropy==2.0.2
attrs==19.3.0
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
blis==0.4.1
bokeh==0.12.10
boto==2.48.0
boto3==1.11.4
botocore==1.14.4
Bottleneck==1.2.1
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.10.0
chardet==3.0.4
chart-studio==1.0.0
chatterbot-corpus==1.2.0
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
colorlover==0.3.0
comtypes==1.1.2
conda==4.5.13
conda-build==3.0.27
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.8
cufflinks==0.17.0
cycler==0.10.0
cymem==2.0.3
Cython==0.26.1
cytoolz==0.8.2
dask==2.9.2
datashape==0.5.4
decorator==4.1.2
distlib==0.2.5
distributed==1.19.1
distro==1.4.0
Django==3.0.3
django-crispy-forms==1.9.0
dlib==19.7.0
dnspython==1.16.0
docutils==0.14
docx==0.2.4
docx2txt==0.8
en-core-web-sm==2.2.5
entrypoints==0.2.3
et-xmlfile==1.0.1
exception==0.1.0
face-recognition==1.2.3
face-recognition-models==0.3.0
fastcache==1.0.2
filelock==2.0.12
firebase-admin==3.2.1
Flask==1.1.1
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.4.1
future==0.18.2
gast==0.2.2
gensim==3.8.1
gevent==1.2.2
glob2==0.5
google==2.0.3
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.1.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-cloud-translate==2.0.0
google-pasta==0.1.8
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
greenlet==0.4.12
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.7.0
heapdict==1.0.0
heapq-max==0.21
html5lib==0.999999999
httplib2==0.17.0
idna==2.9
imageio==2.2.0
imagesize==0.7.1
importlib-metadata==1.5.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.0
Markdown==3.1.1
MarkupSafe==1.0
mathparse==0.1.2
matplotlib==2.1.0
mccabe==0.6.1
menuinst==1.4.10
mistune==0.7.4
mpmath==0.19
msgpack==1.0.0
msgpack-python==0.4.8
multipledispatch==0.4.9
murmurhash==1.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysqlclient==1.4.6
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.4.5
nose==1.3.7
notebook==5.0.0
np-utils==0.5.12.1
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numpy==1.18.1
numpydoc==0.7.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.44
opencv-python==4.1.2.30
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==16.8
pandas==1.0.1
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pdfminer==20191125
pdfminer.six==20200124
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
Pint==0.10.1
pkginfo==1.4.1
plac==0.9.6
playsound==1.2.2
plotly==4.4.1
ply==3.10
preshed==3.0.2
progress==1.3
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.0
py==1.4.34
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.9.4
pycurl==7.43.0.2
pydot==1.4.1
pyflakes==1.6.0
pyforest==1.0.2
Pygments==2.2.0
pylint==1.7.4
pymongo==3.10.1
PyMySQL==0.9.3
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
pypd==1.1.0
pyPdf==1.13
PyPDF2==1.26.0
PyPDF3==1.0.1
pyresparser==1.0.6
pyrsistent==0.15.7
PySocks==1.6.7
pyspark==2.4.4
pytest==3.2.1
python-dateutil==2.8.1
python-util==1.2
python3-utils==0.4.0
pytz==2019.3
PyWavelets==0.5.2
pywin32==221
PyYAML==5.3
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rnn==0.0.0
rope==0.10.5
rsa==4.0
ruamel-yaml==0.11.14
s3transfer==0.3.1
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==1.4.1
seaborn==0.9.0
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.9.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==2.1.0
spacy==2.2.3
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.2.19
sqlparse==0.3.1
srsly==1.0.1
statsmodels==0.8.0
stop-words==2018.7.23
sympy==1.1.1
tables==3.4.2
tabula-py==2.0.4
tabulate==0.8.6
tblib==1.3.2
tensorboard==1.15.0
tensorboardcolab==0.0.22
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
testpath==0.3.1
tflearn==0.3.2
Theano==1.0.4
thinc==7.3.1
toolz==0.8.2
tornado==4.5.2
tqdm==4.11.2
traitlets==4.3.2
typing==3.6.2
unicode==2.7
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
utils==1.0.0
virtualenv==16.7.9
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.6.0
wrapt==1.12.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlutils==2.0.0
xlwings==0.11.4
xlwt==1.3.0
xx-ent-wiki-sm==2.1.0
zict==0.1.3
zipp==2.2.0
